% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComponentsNumber.R
\name{Imputation_Data_PLSreg}
\alias{Imputation_Data_PLSreg}
\title{Title}
\usage{
Imputation_Data_PLSreg(dataX, datay, Ncomp, NK, imputMethod, k = 5, m = 5)
}
\arguments{
\item{dataX}{explicative data with NA's}

\item{datay}{explained data}

\item{Ncomp}{number of components}

\item{k}{number of the nearest neighbors only kNN Imputation method}

\item{m}{number of imputations only for MICE Imputation method}
}
\value{

}
\description{
Title
}
